分開計算維度方法

+++ database/migrations/2020_04_24_032317_create_vacancy_categories_table.php
@@ -16,7 +16,7 @@ class CreateVacancyCategoriesTable extends Migration
         Schema::create('vacancy_categories', function (Blueprint $table) {
             $table->increments('id');
             $table->string('vacancy_category',20);
-            $table->float('weight')->nullable();
+            $table->integer('weight')->nullable();
             $table->timestamps();
         });
     }
